# Speed up processing 'fixit.txt' by removing duplicate entries